Remove redundant same-type casts on nested struct members (3 units)

+++ src/main/dll/dll_00CA_mediumbasket.c
@@ -1281,7 +1281,7 @@ int mediumbasket_stateHandlerB01(int* obj, GroundBaddieState* state)
 {
     extern int* gPlayerInterface; /* #57 */
     GroundBaddieState* sub = ((GameObject*)obj)->extra;
-    if ((s8)state->baddie.hitPoints < 1) return 3;
+    if (state->baddie.hitPoints < 1) return 3;
     if ((s8)state->baddie.moveDone != 0)
     {
         if (state->baddie.controlMode == 12)
